Tani Administratori nuk mundet me delete veten per shkak te logjikes.Mbas regjistrimit redirect ne loginform

+++ delete.php
@@ -1,15 +1,18 @@
 <?php
-
-$useremail = $_GET['id'];
+session_start();
 include_once './userRepository.php';
 
 
+if (!isset($_SESSION['user_role']) || $_SESSION['user_role'] !== 'admin') {
+    header("Location: index.php");
+    exit();
+}
+
 
 $userRepository = new UserRepository();
-
 $userRepository->deleteUser($useremail);
 
-header("location:Administrator.php");
 
-
+header("Location: Administrator.php");
+exit();
 ?>
